Escape html in strings by default
=yield becomes !=yield because the template being yielded to has escaped
its inner html, therefore it is html safe and should not be escaped
(Not doing this means the yielded stuff gets html escaped which makes no
sense)

There was an XSS because the header was displaying the current username
unescaped. The impact was low since the header only displays the
username to the own user, but nevertheless it would be safest to escape
everything.